% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\docType{class}
\name{elections-class}
\alias{elections-class}
\alias{elections}
\title{elections RC Class}
\value{
Nothing.
}
\description{
This class downloads Swedish elections results.
}
\section{Fields}{

\describe{
\item{\code{parliament.}}{Data frame.}

\item{\code{council.}}{Data frame.}

\item{\code{muni.}}{Data frame.}
}}


