% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_citycode.R
\name{getcityid}
\alias{getcityid}
\title{getcityid}
\usage{
getcityid(name)
}
\arguments{
\item{name}{a string, full name of a municipality(may include region-), Case INsensitive}
}
\value{
a string, the id if the municipality you choose
}
\description{
get the id of a municipality in kolada's api
}
\details{
take a name of a sweden municipality, it will print and return the id of that municipality for further usage within kolada's api
}
\examples{

getcityid("linköping")

}
\references{
https://www.kolada.se/
}
