% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcityid_lessdata.R
\name{getcityid_lessdata}
\alias{getcityid_lessdata}
\title{getcityid}
\usage{
getcityid_lessdata(name)
}
\arguments{
\item{name}{a string, full name of a municipality, Case INsensitive}
}
\value{
a string, the id of the municipality you choose
}
\description{
get the id of a municipality in kolada's api with less data GET
}
\details{
take a name of a sweden municipality, it will print and return the id of that municipality for further usage within kolada's api
}
\examples{

getcityid_lessdata("linköping")

}
\references{
https://www.kolada.se/
}
