% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5package-package.R
\docType{package}
\name{lab5package}
\alias{lab5package}
\alias{lab5package-package}
\title{2014 Swedish elections: county results}
\description{
This package contains a reference class that calls the Valmyndigheten API, extracts all the county votes and calculates the percentage of votes for each political party.
See: \code{\link{val2014}}
}
\author{
\strong{Maintainer}: Amer Memic \email{ameme698@student.liu.se}

Authors:
\itemize{
  \item Olof Josefsson \email{olojo524@student.liu.se}
  \item Shahreen Saleem \email{shahreen.saleem111@gmail.com}
}

}
