% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2014-class.R
\docType{class}
\name{val2014}
\alias{val2014}
\title{The \code{val2014} Reference Class}
\value{
A \code{val2014} class generator object.
}
\description{
A Reference Class to calculate the 2014 Swedish election results for all counties. It calls the Valmyndigheten API, reads the semi-colon separated values file, extracts all the votes for each party in every municipality, and calculates the percentage of votes for each political party in every county. These results are stored in the respective county name field.
}
\section{Fields}{

\describe{
\item{\code{Blekinge}}{Blekinge County}

\item{\code{Dalarna}}{Dalarna County}

\item{\code{Gotland}}{Gotland county}

\item{\code{Gavleborg}}{Gävleborg county}

\item{\code{Halland}}{Halland county}

\item{\code{Jamtland}}{Jämtland county}

\item{\code{Jonkoping}}{Jönköping county}

\item{\code{Kalmar}}{Kalmar county}

\item{\code{Kronoberg}}{Kronoberg county}

\item{\code{Norrbotten}}{Norrbotten county}

\item{\code{Skane}}{Skåne county}

\item{\code{Stockholm}}{Stockholm county}

\item{\code{Sodermanland}}{Södermanland county}

\item{\code{Uppsala}}{Uppsala county}

\item{\code{Varmland}}{Värmland county}

\item{\code{Vasterbotten}}{Västerbotten county}

\item{\code{Vasternorrland}}{Västernorrland county}

\item{\code{Vastmanland}}{Västmanland county}

\item{\code{vastergotland}}{Västra Götaland county}

\item{\code{Orebro}}{Örebro county}

\item{\code{Ostergotland}}{Östergötland county}
}}

\section{Methods}{

\describe{
\item{\code{county(county)}}{Returns a named vector with percent values of all political parties for a given county.}

\item{\code{initialize()}}{Calls the Valmyndigheten API to get the 2014 election results of all municipality, then calculating all the election results for every county and storing them in this object.}
}}

\examples{
val = val2014()
print(val$get_county("Blekinge"))
}
