\name{lab5package-package}
\alias{lab5package-package}
\alias{lab5package}
\docType{package}
\title{
\packageTitle{lab5package}
}
\description{
\packageDescription{lab5package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5package}
\packageIndices{lab5package}

}
\author{
\packageAuthor{lab5package}

Maintainer: \packageMaintainer{lab5package}
}

\keyword{ package }

