% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2014-class.R
\docType{class}
\name{val2014}
\alias{val2014}
\title{The \code{val2014} Reference Class}
\value{
A \code{val2014} class generator object.
}
\description{
A Reference Class to calculate the 2014 Swedish election results for all counties. It calls the Valmyndigheten API, reads the semi-colon separated values file, extracts all the votes for each party in every municipality, and calculates the percentage of votes for each political party in every county. These results are stored in the respective county name field.
}
\section{Fields}{

\describe{
\item{\code{Blekinge}}{Blekinge county}

\item{\code{Dalarna}}{Dalarna county}

\item{\code{Gotland}}{Gotland county}

\item{\code{Gavleborg}}{Gavleborg county}

\item{\code{Halland}}{Halland county}

\item{\code{Jamtland}}{Jamtland county}

\item{\code{Jonkoping}}{Jonkoping county}

\item{\code{Kalmar}}{Kalmar county}

\item{\code{Kronoberg}}{Kronoberg county}

\item{\code{Norrbotten}}{Norrbotten county}

\item{\code{Skane}}{Skane county}

\item{\code{Stockholm}}{Stockholm county}

\item{\code{Sodermanland}}{Sodermanland county}

\item{\code{Uppsala}}{Uppsala county}

\item{\code{Varmland}}{Varmland county}

\item{\code{Vasterbotten}}{Vasterbotten county}

\item{\code{Vasternorrland}}{Vasternorrland county}

\item{\code{Vastmanland}}{Vastmanland county}

\item{\code{vastergotland}}{Vastra Gotaland county}

\item{\code{Orebro}}{Orebro county}

\item{\code{Ostergotland}}{Ostergotland county}
}}

\section{Methods}{

\describe{
\item{\code{county(county)}}{Returns a named vector with percent values of all political parties for a given county.}

\item{\code{initialize()}}{Calls the Valmyndigheten API to get the 2014 election results of all municipality, then calculating all the election results for every county and storing them in this object.}
}}

\examples{
val = val2014()
print(val$county("Blekinge"))
}
